requirements.txt
flask
pyaudio
pygame
schedule
python-vlc
pycaw
comtypes
easygui
